requirements.txt
Django==4.2
psycopg2==2.9.9
pillow==10.4
whitenoise==6.7.0
gunicorn==23.0.0
django-environ==0.11.2
djangorestframework==3.15.2
django-cors-headers==4.5.0

